backend/requirements.txt
beautifulsoup4==4.6.3
certifi==2018.10.15
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==2.1.3
django-appconf==1.0.2
django-compressor==2.2
django-cors-headers==2.4.0
django-sass-processor==0.7.2
djangorestframework==3.9.0
djangorestframework-recursive==0.1.2
djangorestframework-xml==1.3.0
djangorestframework-yaml==1.0.3
idna==2.7
itypes==1.1.0
Jinja2==2.10
libsass==0.16.0
MarkupSafe==1.1.0
Pillow==5.3.0
Pygments==2.3.0
pytz==2018.7
PyYAML==3.13
rcssmin==1.0.6
requests==2.20.1
rjsmin==1.0.12
six==1.11.0
uritemplate==3.0.0
urllib3==1.24.1
uwsgi==2.0.17.1
